requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.6
jwt==1.3.1
mysql-connector-python==8.3.0
pony==0.7.17
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.29.0
